requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
PyMySQL==1.1.1
sqlparse==0.5.1
drf-spectacular==0.27.0
flake8==7.1.1